requirements.txt
requests
python-dotenv
discord.py
llama-cpp-python[server]
mysql-connector-python